Modif soutenance

+++ Backend/controllers/sauces.js
@@ -1,4 +1,5 @@
 const Sauces = require('../models/sauces');
+const fs = require('fs');
 
 exports.createSauce = (req, res, next) => {
     const sauceObject = JSON.parse(req.body.sauce);
@@ -26,6 +27,16 @@ exports.getOneSauce = (req, res, next) => {
 }
 
 exports.modifySauce = (req, res, next) => {
+    if (req.file){
+        Sauces.findOne({ _id: req.params.id })
+        .then ((sauce) =>{
+            const filename = sauce.imageUrl.split("/images/")[1];
+            fs.unlink(`images/${filename}`,(error)=>{
+                if (error) console.log(error)
+            }) 
+        })
+        .catch((error) => console.log(error));
+    }
     const sauceObject = req.file ? {
         ...JSON.parse(req.body.sauce),
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
@@ -46,9 +57,20 @@ exports.modifySauce = (req, res, next) => {
 };
 
 exports.deleteSauce = (req, res, next) => {
-    Sauces.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: "Sauce supprimée !" }))
-        .catch(error => res.status(401).json({ error }));
+    Sauces.findOne({ _id: req.params.id })
+        .then(sauce => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({ message: "Non authorisé" })
+            } else {
+                const filename = sauce.imageUrl.split("/images/")[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauces.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: "Sauce supprimée !" }))
+                        .catch(error => res.status(401).json({ error }));
+                })
+
+            }
+        })
 }
 
 exports.getAllSauce = (req, res, next) => {
